requirements.txt
tweepy==4.16.0
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
schedule==1.2.1
